pyproject.toml
[tool.poetry]
name = "test-cybellum"
version = "0.1.0"
description = "small flask rest api"
authors = ["Sergey <suprun.sergey89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
sqlalchemy = "^2.0.28"
pydantic = "^2.6.4"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"
black = "^24.2.0"
flake8 = "^7.0.0"
pyjwt = "^2.8.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
